[test] Add tests to xor.go to ensure the accuracy of "delta of delta" compact algorithm and "XOR" compact algorithm #6885 Signed-off-by: fuling <fuling.lgz@alibaba-inc.com>
